enhancement-967: divided into model and view
