+++ js/mcqView.js
@@ -1,0 +1,186 @@
+define([
+    'coreViews/questionView'
+], function(QuestionView) {
+
+    var McqView = QuestionView.extend({
+
+        events: {
+            'focus .mcq-item input':'onItemFocus',
+            'blur .mcq-item input':'onItemBlur',
+            'change .mcq-item input':'onItemSelected',
+            'keyup .mcq-item input':'onKeyPress'
+        },
+
+        resetQuestionOnRevisit: function() {
+            this.setAllItemsEnabled(true);
+            this.resetQuestion();
+        },
+
+        setupQuestion: function() {
+            this.model.setupRandomisation();
+        },
+
+        disableQuestion: function() {
+            this.setAllItemsEnabled(false);
+        },
+
+        enableQuestion: function() {
+            this.setAllItemsEnabled(true);
+        },
+
+        setAllItemsEnabled: function(isEnabled) {
+            _.each(this.model.get('_items'), function(item, index){
+                var $itemLabel = this.$('label').eq(index);
+                var $itemInput = this.$('input').eq(index);
+
+                if (isEnabled) {
+                    $itemLabel.removeClass('disabled');
+                    $itemInput.prop('disabled', false);
+                } else {
+                    $itemLabel.addClass('disabled');
+                    $itemInput.prop('disabled', true);
+                }
+            }, this);
+        },
+
+        onQuestionRendered: function() {
+            this.setReadyStatus();
+            if (!this.model.get("_isSubmitted")) return;
+            this.showMarking();
+        },
+
+        onKeyPress: function(event) {
+            if (event.which === 13) { //<ENTER> keypress
+                this.onItemSelected(event);
+            }
+        },
+
+        onItemFocus: function(event) {
+            if(this.model.get('_isEnabled') && !this.model.get('_isSubmitted')){
+                $("label[for='"+$(event.currentTarget).attr('id')+"']").addClass('highlighted');
+            }
+        },
+        
+        onItemBlur: function(event) {
+            $("label[for='"+$(event.currentTarget).attr('id')+"']").removeClass('highlighted');
+        },
+
+        onItemSelected: function(event) {
+            if(this.model.get('_isEnabled') && !this.model.get('_isSubmitted')){
+                var selectedItemObject = this.model.get('_items')[$(event.currentTarget).parent('.component-item').index()];
+                this.toggleItemSelected(selectedItemObject, event);
+            }
+        },
+
+        toggleItemSelected:function(item, clickEvent) {
+            var selectedItems = this.model.get('_selectedItems');
+            var itemIndex = _.indexOf(this.model.get('_items'), item),
+                $itemLabel = this.$('label').eq(itemIndex),
+                $itemInput = this.$('input').eq(itemIndex),
+                selected = !$itemLabel.hasClass('selected');
+            
+                if(selected) {
+                    if(this.model.get('_selectable') === 1){
+                        this.$('label').removeClass('selected');
+                        this.$('input').prop('checked', false);
+                        this.deselectAllItems();
+                        selectedItems[0] = item;
+                    } else if(selectedItems.length < this.model.get('_selectable')) {
+                     selectedItems.push(item);
+                 } else {
+                    clickEvent.preventDefault();
+                    return;
+                }
+                $itemLabel.addClass('selected');
+                $itemLabel.a11y_selected(true);
+            } else {
+                selectedItems.splice(_.indexOf(selectedItems, item), 1);
+                $itemLabel.removeClass('selected');
+                $itemLabel.a11y_selected(false);
+            }
+            $itemInput.prop('checked', selected);
+            item._isSelected = selected;
+            this.model.set('_selectedItems', selectedItems);
+        },
+
+        // Blank method to add functionality for when the user cannot submit
+        // Could be used for a popup or explanation dialog/hint
+        onCannotSubmit: function() {},
+
+        // This is important and should give the user feedback on how they answered the question
+        // Normally done through ticks and crosses by adding classes
+        showMarking: function() {
+            _.each(this.model.get('_items'), function(item, i) {
+                var $item = this.$('.component-item').eq(i);
+                $item.removeClass('correct incorrect').addClass(item._isCorrect ? 'correct' : 'incorrect');
+            }, this);
+        },
+
+        // Used by the question view to reset the look and feel of the component.
+        resetQuestion: function() {
+            this.deselectAllItems();
+            this.resetItems();
+        },
+
+        deselectAllItems: function() {
+            this.$el.a11y_selected(false);
+            this.model.deselectAllItems();
+        },
+
+        resetItems: function() {
+            this.$('.component-item label').removeClass('selected');
+            this.$('.component-item').removeClass('correct incorrect');
+            this.$('input').prop('checked', false);
+            this.model.resetItems();
+        },
+
+        showCorrectAnswer: function() {
+            _.each(this.model.get('_items'), function(item, index) {
+                this.setOptionSelected(index, item._shouldBeSelected);
+            }, this);
+        },
+
+        setOptionSelected:function(index, selected) {
+            var $itemLabel = this.$('label').eq(index);
+            var $itemInput = this.$('input').eq(index);
+            if (selected) {
+                $itemLabel.addClass('selected');
+                $itemInput.prop('checked', true);
+            } else {
+                $itemLabel.removeClass('selected');
+                $itemInput.prop('checked', false);
+            }
+        },
+
+        hideCorrectAnswer: function() {
+            _.each(this.model.get('_items'), function(item, index) {
+                this.setOptionSelected(index, this.model.get('_userAnswer')[item._index]);
+            }, this);
+        },
+
+        /**
+        * used by adapt-contrib-spoor to get the user's answers in the format required by the cmi.interactions.n.student_response data field
+        * returns the user's answers as a string in the format "1,5,2"
+        */
+        getResponse:function() {
+            var selected = _.where(this.model.get('_items'), {'_isSelected':true});
+            var selectedIndexes = _.pluck(selected, '_index');
+            // indexes are 0-based, we need them to be 1-based for cmi.interactions
+            for (var i = 0, count = selectedIndexes.length; i < count; i++) {
+                selectedIndexes[i]++;
+            }
+            return selectedIndexes.join(',');
+        },
+
+        /**
+        * used by adapt-contrib-spoor to get the type of this question in the format required by the cmi.interactions.n.type data field
+        */
+        getResponseType:function() {
+            return "choice";
+        }
+
+    });
+
+    return McqView;
+
+});
